heat exchanger updates
gave the heat exchanger model script a way to run as main and a way to run in the shell as well as default inputs for testing.

Moved the experimental design section ahead of the results section

Expanded on the test case section with heat exchanger model information

Fixed formatting of printed values in the heat exchanger model.